refactor: reorganize provider imports
