+++ internal/providers/weather/gismeteo/response.go
